nutriflex-core/requirements.txt
crewai==0.28.8
streamlit==1.28.1
openai>=1.13.3
python-dotenv==1.0.0
anthropic==0.7.7